rename example pages
